04 Django/django/10_drf/代码/day10/requirement.txt
Django==2.2
djangorestframework==3.11.0
docopt==0.6.2
mysqlclient==1.4.6
psm==0.1.2
pytz==2019.3
sqlparse==0.3.1
